Pin imageio to version that doesn't break moviepy
Recently imageio has been upgraded and now throws an error
if imageio-ffmpeg isn't installed, even when you install it moviepy is
still broken, for the time being we need to depend on imageio<2.5.0 until
they release a version with less insane behaviour or moviepy is fixed.

This was reported in issue https://github.com/epic-kitchens/epic-lib/issues/69
and is currently breaking all CI tests.